Row and column add buttons refactored
